Noop: add a development provider for exploration.
During the exploration of various clouds, most developers try to boot up
a system to poke around. This provides a bioler-plate Afterburn provider
called 'noop' that does absolutely nothing. The original idea came with
a wire-framing of AzureStack and needing a base to start with. Finally,
the noop provider can provide a "template" for getting started with
providers.

Adds:
- 'noop' provider
- 'AzureStack' provider which uses the 'noop' provider

Signed-off-by: Ben Howard <ben.howard@redhat.com>